build: resolve compile error on 32-bit
exch/ews/requests.cpp:706:41: error: no matching function for call to
'min(std::vector<std::__cxx11::basic_string<char> >::size_type, long
unsigned int)'
   aliases.resize(min(aliases.size(), 3ul));

+++ exch/ews/requests.cpp
@@ -703,7 +703,7 @@ void process(mResolveNamesRequest&& request, XMLElement* response, const EWSCont
 	if(!ctx.plugin.mysql.get_user_aliases(request.UnresolvedEntry.c_str(), aliases))
 		throw DispatchError(E3068);
 	if (aliases.size() > 0) {
-		aliases.resize(min(aliases.size(), 3ul));
+		aliases.resize(min(aliases.size(), size_t(3)));
 		cnt.EmailAddresses.emplace().reserve(aliases.size());
 		size_t index = 0;
 		for (auto& alias : aliases)
